+++ opal-r/src/main/java/org/obiba/opal/web/r/RServiceClusterResource.java
@@ -15,9 +15,11 @@ import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.UriBuilder;
 import jakarta.ws.rs.core.UriInfo;
+import org.obiba.opal.core.service.OpalGeneralConfigService;
 import org.obiba.opal.web.model.OpalR;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
@@ -32,6 +34,9 @@ import java.util.List;
 public class RServiceClusterResource extends AbstractRServiceResource {
 
   private static final Logger log = LoggerFactory.getLogger(RServiceClusterResource.class);
+
+  @Autowired
+  private OpalGeneralConfigService opalGeneralConfigService;
 
   @GET
   public OpalR.RServerClusterDto getCluster(@PathParam("cname") String clusterName) {
@@ -59,6 +64,9 @@ public class RServiceClusterResource extends AbstractRServiceResource {
   @PUT
   @Path("/packages")
   public Response updateAllPackages(@PathParam("cname") String clusterName) {
+    if (!opalGeneralConfigService.getConfig().isAllowRPackageManagement())
+      return Response.status(Response.Status.FORBIDDEN).build();
+
     try {
       rPackageHelper.updateAllCRANPackages(rServerManagerService.getRServerCluster(clusterName));
     } catch (Exception e) {
@@ -71,6 +79,9 @@ public class RServiceClusterResource extends AbstractRServiceResource {
   @Path("/packages")
   public Response installPackage(@Context UriInfo uriInfo, @PathParam("cname") String clusterName, @QueryParam("name") String pkgName,
                                  @QueryParam("ref") String ref, @QueryParam("manager") String manager) {
+    if (!opalGeneralConfigService.getConfig().isAllowRPackageManagement())
+      return Response.status(Response.Status.FORBIDDEN).build();
+
     rPackageHelper.installPackage(rServerManagerService.getRServerCluster(clusterName), pkgName, ref, manager);
     UriBuilder ub = uriInfo.getBaseUriBuilder().path(RServicePackageResource.class);
     return Response.created(ub.build(pkgName)).build();
@@ -80,6 +91,9 @@ public class RServiceClusterResource extends AbstractRServiceResource {
   @Path("/packages")
   public Response deletePackages(@PathParam("cname") String clusterName, 
                                  @QueryParam("name") List<String> pkgNames) {
+    if (!opalGeneralConfigService.getConfig().isAllowRPackageManagement())
+      return Response.status(Response.Status.FORBIDDEN).build();
+
     if (pkgNames != null) {
       pkgNames.forEach(pkgName -> deletePackage(clusterName, pkgName));
     }
@@ -104,6 +118,9 @@ public class RServiceClusterResource extends AbstractRServiceResource {
   @DELETE
   @Path("/package/{pname}")
   public Response deletePackage(@PathParam("cname") String clusterName, @PathParam("pname") String pkgName) {
+    if (!opalGeneralConfigService.getConfig().isAllowRPackageManagement())
+      return Response.status(Response.Status.FORBIDDEN).build();
+
     try {
       rPackageHelper.removePackage(rServerManagerService.getRServerCluster(clusterName), pkgName);
     } catch (Exception e) {
